requirements.txt
gunicorn==20.0.4
certifi>=2022.12.07
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.5
pymysql
flask-sqlalchemy
websocket_client==1.6.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
munch==2.5.0
requests>=2.31.0
six==1.16.0
urllib3==1.26.17
Werkzeug>=2.2.3
python-dotenv==0.20.0
uvicorn
twilio
python-decouple
python-multipart
pyngrok